apps/migration_viewer/requirements.txt
streamlit>=1.28
psycopg2-binary>=2.9
pandas>=2.0
python-dotenv>=1.0
sqlalchemy>=2.0
